wind-turbine-dashboard/requirements.txt
paho-mqtt==1.5.1
influxdb-client==1.23.0
